updates for AppMenu
